+++ src/plonemeeting/portal/core/content/subscriber.py
@@ -9,7 +9,7 @@ from zope.i18n import translate
 from plonemeeting.portal.core import _
 from plonemeeting.portal.core.interfaces import IItemsFolder
 from plonemeeting.portal.core.interfaces import IMeetingsFolder
-from plonemeeting.portal.core.utils import create_faceted_folder
+from plonemeeting.portal.core.utils import create_faceted_folder, set_constrain_types
 from plonemeeting.portal.core.utils import format_institution_managers_group_id
 
 
@@ -29,6 +29,7 @@ def handle_institution_creation(obj, event):
     )
     alsoProvides(meetings, IMeetingsFolder)
     IFacetedLayout(meetings).update_layout("faceted-preview-meeting")
+    set_constrain_types(meetings, [])
 
     # Create items faceted folder
     items = create_faceted_folder(
@@ -36,6 +37,7 @@ def handle_institution_creation(obj, event):
     )
     alsoProvides(items, IItemsFolder)
     IFacetedLayout(items).update_layout("faceted-preview-meeting-items")
+    set_constrain_types(items, [])
 
     request = getRequest()
     if request:  # Request can be `None` during test setup