+++ CHANGES.rst
@@ -6,6 +6,11 @@ Changelog
 ------------------
 
 - Query categories from iA.Delib to populate vocabularies only when loading the edit form.
+- Fixed long representative value ws not used.
+  [odelaere]
+- Fixed error while compiling rules.xml by institution manager.
+  [gbastien]
+- Fix type constraints on Folder content type and faceted folders.
   [odelaere]